Refactor shared models to their own files and package
