+++ core/command_dispatcher.py
@@ -4,35 +4,17 @@ import os
 import importlib.util
 import logging
 
-from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from core.event_bus import EventBus
 from interfaces.bot_command import BotCommand
+from models.command import CommandData
 from services.meshtastic_service import TEXT_MESSAGE_TOPIC
 
 if TYPE_CHECKING:
     from services.meshtastic_service import TextPacket
 
 COMMANDS_DIR = "./commands"
-
-
-@dataclass
-class CommandData:
-    """
-    Data class to hold command information.
-    """
-    __slots__ = ['sender_id', 'receiver_id', 'parameters', 'raw_message',
-                 'channel', 'rx_time', 'rx_snr', 'hops_away', 'via_mqtt']
-    sender_id: str
-    receiver_id: str
-    parameters: list[str] | None
-    raw_message: str
-    channel: int | None
-    rx_time: int
-    rx_snr: float | None
-    hops_away: int | None
-    via_mqtt: bool
 
 
 class CommandDispatcher:
