+++ models/location.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class GpsLocation:
+    """
+    Represents a GPS location with latitude and longitude.
+    """
+    __slots__ = ['latitude', 'longitude']
+    latitude: float
+    longitude: float
