+++ modules/node_search.py
@@ -1,8 +1,8 @@
 from datetime import datetime
 
-from core.command_dispatcher import CommandData
-from core.database import NodeInfo
 from interfaces.bot_module import BotModule
+from models.command import CommandData
+from models.node import NodeInfo
 from utils.geo_utils import get_city_state_offline, get_lat_lon_from_string, calculate_distance
