+++ modules/weather_conditions.py
@@ -1,5 +1,7 @@
-from core.command_dispatcher import CommandData
 from interfaces.bot_module import BotModule
+from models.command import CommandData
+from models.location import GpsLocation
+from models.weather import WeatherConditions
 from services.positionstack_geocode_service import PositionstackGeocodeService
 from services.nws_weather_service import NwsWeatherService
 
@@ -41,7 +43,7 @@ class WeatherConditions(BotModule):
             self.mesh_service.send_reply("You must provide a location.", data)
             return
         query = ' '.join(arguments)
-        coords = self.geo_service.get_coords(query)
+        coords: GpsLocation | None = self.geo_service.get_coords(query)
         if coords is None:
             self.mesh_service.send_reply(
                 "Unable to identify the location for your query.", data)
@@ -51,7 +53,7 @@ class WeatherConditions(BotModule):
             self.mesh_service.send_reply(
                 "Unable to identify the location for your query.", data)
             return
-        conditions = self.api_service.get_conditions(zone)
+        conditions: WeatherConditions | None = self.api_service.get_conditions(zone)
         if conditions is None:
             self.mesh_service.send_reply(
                 "Unable to lookup the conditions for that location.", data)
