+++ modules/weather_forecast.py
@@ -1,5 +1,7 @@
-from core.command_dispatcher import CommandData
 from interfaces.bot_module import BotModule
+from models.command import CommandData
+from models.location import GpsLocation
+from models.weather import WeatherForecast
 from services.positionstack_geocode_service import PositionstackGeocodeService
 from services.nws_weather_service import NwsWeatherService
 
@@ -41,7 +43,7 @@ class WeatherForecast(BotModule):
             self.mesh_service.send_reply("You must provide a location.", data)
             return
         query = ' '.join(arguments)
-        coords = self.geo_service.get_coords(query)
+        coords: GpsLocation | None = self.geo_service.get_coords(query)
         if coords is None:
             self.mesh_service.send_reply(
                 "Unable to identify the location for your query.", data)
@@ -51,7 +53,8 @@ class WeatherForecast(BotModule):
             self.mesh_service.send_reply(
                 "Unable to identify the location for your query.", data)
             return
-        forecasts = self.api_service.get_forecasts(zone)
+        forecasts: list[WeatherForecast] | None = self.api_service.get_forecasts(
+            zone)
         if forecasts is None or len(forecasts) <= 0:
             self.mesh_service.send_reply(
                 "Unable to lookup the conditions for that location.", data)
